<Done> Create Fucntion to add Sell Contracts

+++ pkg/models/contracts.go
@@ -6,7 +6,7 @@ type Contract struct {
 	UserID      string  `json:"user_id"`
 	RequestType string  `json:"request_type"`
 	OrderType   string  `json:"order_type"`
-	Price       float64 `json:"price" binding:"required,gt=0"`
-	Quantity    int     `json:"quantity" binding:"required,gte=1"`
+	Price       float32 `json:"price" binding:"required,gt=0"`
+	Quantity    int64   `json:"quantity" binding:"required,gte=1"`
 	Timestamp   int64   `json:"timestamp"`
 }
